requirements.txt
albucore==0.0.17
albumentations==1.4.18
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.13.5
black==24.8.0
Brotli==1.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
coloredlogs==15.0.1
Deprecated==1.2.18
dj-database-url==2.3.0
Django==4.2.20
django-environ==0.11.2
easydict==1.13
environ==1.0
eval-type-backport==0.2.2
filelock==3.16.1
flatbuffers==25.9.23
fpdf==1.7.2
fsspec==2025.3.0
gdown==5.2.0
gunicorn==23.0.0
hf-xet==1.1.10
huggingface-hub==0.34.5
humanfriendly==10.0
idna==3.10
imageio==2.35.1
img2pdf==0.6.1
importlib-metadata==8.5.0
importlib-resources==6.4.5
jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kornia==0.7.3
kornia-rs==0.1.9
lazy-loader==0.4
llvmlite==0.41.1
lxml==6.0.2
MarkupSafe==2.1.5
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
onnxruntime==1.16.3
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==24.2
pathspec==0.12.1
pikepdf==9.2.1
pillow==10.4.0
pkgutil-resolve-name==1.3.10
platformdirs==4.3.6
pooch==1.8.2
protobuf==5.29.5
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic-core==2.27.2
pymatting==1.1.14
PySocks==1.7.1
python-decouple==3.8
python-dotenv==1.0.1
pytz==2025.1
PyWavelets==1.4.1
PyYAML==6.0.2
referencing==0.35.1
rembg==2.0.61
requests==2.32.4
rpds-py==0.20.1
safetensors==0.5.3
scikit-image==0.21.0
scipy==1.10.1
soupsieve==2.7
sqlparse==0.5.3
sympy==1.13.3
tifffile==2023.7.10
timm==1.0.19
tomli==2.2.1
torch==2.4.1
torchaudio==2.4.1+cu118
torchvision==0.19.1
tqdm==4.67.1
transparent-background==1.3.4
triton==3.0.0
typing-extensions==4.12.2
urllib3==2.2.3
wget==3.2
whitenoise==6.7.0
wrapt==1.17.3
zipp==3.20.2
